separating different types of model architectures
