+++ transformer/__init__.py
@@ -1,1 +1,4 @@
-from .transformer import GPT, DataLoader, DDPConfig, GPTConfig, MLP, Block, CausalSelfAttention
+from .gpt_no_layernorm import GPT_noLN
+from .last_token_gpt import LastTokenGPT
+from .transformer import (GPT, MLP, Block, CausalSelfAttention, DataLoader,
+                          DDPConfig, GPTConfig)
