+++ transformer/last_token_gpt.py
@@ -1,16 +1,8 @@
-import os
-import sys
-from dataclasses import dataclass
-from datetime import timedelta
-
 import torch
-import torch.distributed as dist
 import torch.nn as nn
 from torch.nn import functional as F
 
-from utils.file_management import setup_project_path
 from transformer import GPT, MLP, Block, CausalSelfAttention
-
 
 seed = 200
 torch.manual_seed(seed)
@@ -235,49 +227,6 @@ class LastTokenGPT(GPT):
         # Add sequence dimension back for consistency with training
         logits = logits.unsqueeze(1)  # Shape: (B, 1, vocab_size)
         
-        loss = self.calculate_loss(logits, targets, **kwargs)
-
-        return logits, loss
-
-
-class Block_noLN(nn.Module):
-    """Transformer block with attention and MLP"""
-    def __init__(self, config):
-        super().__init__()
-        self.attn = CausalSelfAttention(config)
-        self.mlp = MLP(config)
-
-    def forward(self, x):
-        x = x + self.attn(x)
-        x = x + self.mlp(x)
-        return x
-
-
-class GPT_noLN(GPT):
-    """GPT-style transformer model for sequence prediction"""
-    def __init__(self, config):
-        super().__init__(config)
-        self.transformer = nn.ModuleDict({
-            'wte': nn.Embedding(config.vocab_size, config.n_embd),
-            'wpe': nn.Embedding(config.block_size, config.n_embd),
-            'h': nn.ModuleList([Block_noLN(config) for _ in range(config.n_layer)]),
-        })
-        self.lm_head = nn.Linear(config.n_embd, config.vocab_size, bias=False)
-        self.lm_head.weight = self.transformer.wte.weight  # Weight tying
-        self.apply(self._init_weights)
-
-    def forward(self, idx, targets=None, **kwargs):
-        """Forward pass through the model"""
-        B, T = idx.size()
-        assert T <= self.config.block_size, f"Sequence length {T} exceeds block size {self.config.block_size}"
-
-        pos = torch.arange(0, T, dtype=torch.long, device=idx.device)
-        x = self.transformer.wte(idx) + self.transformer.wpe(pos)
-
-        for block in self.transformer.h:
-            x = block(x)
-
-        logits = self.lm_head(x)
         loss = self.calculate_loss(logits, targets, **kwargs)
 
         return logits, loss
